requirements.txt
db-sqlite3==0.0.1
Django==2.1.1
djangorestframework==3.9.2
python-decouple==3.1